Automated much of the user work during manual sweep. Takes about 10 mins per FL now.
Added phase curve to timestream (with low alpha) for reference. Added a check that the phase and intensities peaks are in roughly the same place. When the timestream is marked as a double the peak location moves to the bigger phase dip. XY title tells the user how to set the good region and once it is set. Pixels outside this region are automatically labelled bad during fast forward.